// src/api/generated/model/todoDtoCat.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type TodoDtoCat = typeof TodoDtoCat[keyof typeof TodoDtoCat];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TodoDtoCat = {
  tech: 'tech',
  freizeit: 'freizeit',
} as const;